+++ modules/canvaskit/font.js
@@ -128,21 +128,6 @@ CanvasKit._extraInitializations.push(function() {
     return fm;
   };
 
-  // TODO(kjlubick) Remove this deprecated API
-  CanvasKit.FontMgr.prototype.MakeTypefaceFromData = function(fontData) {
-    var data = new Uint8Array(fontData);
-
-    var fptr = copy1dArray(data, 'HEAPU8');
-    var font = this._makeTypefaceFromData(fptr, data.byteLength);
-    if (!font) {
-      Debug('Could not decode font data');
-      // We do not need to free the data since the C++ will do that for us
-      // when the font is deleted (or fails to decode);
-      return null;
-    }
-    return font;
-  };
-
   CanvasKit.Typeface.MakeFreeTypeFaceFromData = function(fontData) {
     var data = new Uint8Array(fontData);