+++ modules/canvaskit/CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Removed
+ - `FontMgr.MakeTypefaceFromData` and `FontMgr.RefDefault` have been removed in favor of
+   `Typeface.MakeFreeTypeFaceFromData`
+
 ### Type Changes (index.d.ts)
  - `Surface.requestAnimationFrame` and `Surface.drawOnce` are properly documented.
  - Fixed typo in TextStyle (decrationStyle => decorationStyle)